requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==2.0.1
BaseHash==3.0.10
beautifulsoup4==4.11.2
bootstrap-admin==0.4.4
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-admin-charts==1.2.4
django-admin-tools==0.9.2
django-admin-tools-stats==0.9.0
django-bower==5.2.0
django-crispy-forms==1.14.0
django-data-browser==4.0.12
django-easy-audit==1.3.3
django-environ==0.9.0
django-import-export==3.0.2
django-jsonfield==1.4.1
django-ledger==0.5.0.5
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-qsstats-magic==1.1.0
django-report-builder==6.4.2
django-slick-reporting==0.6.5
django-treebeard==4.5.1
djangorestframework==3.14.0
djcacheutils==3.0.0
et-xmlfile==1.1.0
Faker==8.16.0
fonttools==4.38.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
Jinja2==3.1.2
jsonschema==3.2.0
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
ofxtools==0.9.4
openpyxl==3.0.10
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.3
PyYAML==6.0
simplejson==3.18.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.3.0
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
Unidecode==1.3.6
weasyprint==57.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.2
